Add getLoginUrl and getLineUserData methods.
These methods allow to generate the Line Login URL, and the second
method will get the user data and return an object with the exact
fields to append to any user collection, in order for that user to
be able to login into the app.
These methods should be used when the application has required fields
(for example roles) that the OAuth package does not insert automatically.